Fix CI formatting and middleware test size gate
